Rag_Code/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
chromadb>=0.4.22
neo4j>=5.15.0
google-generativeai==0.3.2
pydantic==2.5.3
sqlalchemy==2.0.23
PyPDF2==3.0.1
pypdf==3.17.4
beautifulsoup4==4.12.2
requests==2.31.0
sentence-transformers==2.2.2
python-dotenv==1.0.0
numpy==1.26.2
pandas==2.1.4
